requirements.txt
pillow==11.2.1
transformers==4.51.3
torch==2.7.0
fastapi==0.112.2
bcrypt==4.0.1
cryptography==43.0.0
python-jose==3.3.0
SQLAlchemy==2.0.32
uvicorn==0.30.6
passlib==1.7.4
langchain==0.2.15
python-dotenv==1.0.1
google-generativeai==0.7.2
langchain-google-genai==1.0.10
httpx==1.4.0
streamlit==1.24.0